+++ pkg/netflow/testutil/testutil.go
@@ -1,9 +1,9 @@
 // Unless explicitly stated otherwise all files in this repository are licensed
 // under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2022-present Datadog, Inc.
+// Copyright 2023-present Datadog, Inc.
 
-package goflowlib
+package testutil
 
 import (
 	"fmt"
